Techdebt: remove the vendored `boto` package
`moto/packages/boto` was added when the direct `boto` dependency was
dropped, with the stated next step of folding the remaining classes into
their moto subclasses so the directory could be deleted entirely.

* `blockdevicemapping.py` had no boto inheritance left, so it moves to
  `moto/ec2/models/block_device.py` unchanged, minus the unused
  `EBSBlockDeviceType` alias.
* `InstancePlacement` and `Reservation` are now defined in
  `moto/ec2/models/instances.py`, and `Instance` no longer inherits from
  the vendored `boto` `Instance`. The inherited attributes that are
  actually load-bearing -- `_placement`, `hypervisor`, `product_codes`,
  `kernel` and `ramdisk` -- are seeded in `Instance.__init__` instead.
  The other ~20 were dead, including `tags`, which was already shadowed
  by `TaggedEC2Resource.tag_list` during serialization.
* `instancetype.py` had no importers at all.

Three call sites were reading attributes that the vendored base class
only ever set to a constant:

* `autoscaling` assigned the `InstancePlacement` object to
  `availability_zone` and relied on its `__repr__` being stringified
  during serialization; it now reads `.placement.zone`.
* `get_launch_template_data` read `instance.monitored`, always `False`.
* `emr.run_instances` rebound its loop variable from the EC2 `Instance`
  to the EMR `Instance`. `Reservation.instances` was typed `Any` before,
  so mypy could not see it.

Responses are unchanged: `DescribeInstances` (including the
`availability-zone` and `product-code` filters), all thirteen
`DescribeInstanceAttribute` attributes, `GetLaunchTemplateData`,
`DescribeImages`, `DescribeAutoScalingGroups` and
`DescribeLaunchConfigurations` were diffed before and after.
